Store do_package in sstate-cache
Split do_package into two tasks:
- do_package: capture of packaging files
- do_package_split: division of packaging files and generation of
                    the pkgdata and dependency on pkgdata of
                    package dependencies

Dividing the task into two allows us to capture the packaged contents
without any of the recipe specific pkgdata - making a clean base for the
sstate-cache contents. The pkgdata generated on top in do_packages_split
can then be applied on top of the cached data.

This is needed to avoid re-compiling recipes when the pkgdata changes
(such as the PR of the recipe or the pkgdata of a dependency) and
instead re-use the compilation results. This saves a lot of time as
re-packaging is often less costly than re-compiling and re-packaging.

The alternative approach to this which is to sstate-cache do_install instead
is not used here because it leads to other complicated problems I could
not solve.

Signed-off-by: Michael Ho <michael.ho@ieee.org>